move full op to createon.py test=develop
