+++ python/paddle/__init__.py
@@ -51,7 +51,7 @@ import paddle.sysconfig
 # from .tensor.creation import zeros_like   #DEFINE_ALIAS
 # from .tensor.creation import arrange   #DEFINE_ALIAS
 # from .tensor.creation import eye   #DEFINE_ALIAS
-# from .tensor.creation import full   #DEFINE_ALIAS
+from .tensor.creation import full  #DEFINE_ALIAS
 # from .tensor.creation import linspace   #DEFINE_ALIAS
 # from .tensor.creation import full_like   #DEFINE_ALIAS
 # from .tensor.creation import triu   #DEFINE_ALIAS
