+++ python/paddle/tensor/__init__.py
@@ -29,7 +29,7 @@
 # from .creation import zeros_like   #DEFINE_ALIAS
 # from .creation import arrange   #DEFINE_ALIAS
 # from .creation import eye   #DEFINE_ALIAS
-# from .creation import full   #DEFINE_ALIAS
+from .creation import full  #DEFINE_ALIAS
 # from .creation import linspace   #DEFINE_ALIAS
 # from .creation import full_like   #DEFINE_ALIAS
 # from .creation import triu   #DEFINE_ALIAS
